+++ springTest/src/main/java/com/chandana/customer/CustomerRowMapper.java
@@ -12,7 +12,7 @@ public class CustomerRowMapper implements RowMapper<Customer>{
     @Override
     public Customer mapRow(ResultSet rs, int rowNum) throws SQLException {
         return new Customer(
-            rs.getLong("id"),
+            rs.getInt("id"),
             rs.getString("name"),
             rs.getString("email"),
             rs.getInt("age")
